+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/brokenhieararchy/BrokenHierarchyInterface.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.java.generics.resolver.support.brokenhieararchy;
+
+import java.util.concurrent.Callable;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 11.02.2015
+ */
+public interface BrokenHierarchyInterface<T extends Callable, K> {
+}