+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/brokenhieararchy/BrokenHierarchyRoot.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.java.generics.resolver.support.brokenhieararchy;
+
+/**
+ * Generic not set on base class and implemented interface.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 11.02.2015
+ */
+public class BrokenHierarchyRoot extends BrokenHierarchyBase implements BrokenHierarchyInterface {
+}